@@ -6,7 +6,8 @@
   "scripts": {
     "build": "xcodebuild -project ios/CalendarsExample.xcodeproj build",
     "e2e": "node ./scripts/test-e2e.js --release",
-    "test": "jasmine src/*.spec.js && npm run lint && npm run e2e",
+    "test": "npm run lint && npm run unit && npm run e2e",
+    "unit": "jest",
     "lint": "eslint src/ example/src",
     "release": "node ./scripts/release.js",
     "start": "react-native start",
@@ -36,17 +37,40 @@
     "xdate": "^0.8.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.12.9",
+    "@babel/runtime": "^7.12.5",
+    "@welldone-software/why-did-you-render": "^6.0.3",
     "babel-eslint": "^7.2.3",
+    "babel-jest": "^26.6.3",
     "detox": "^16.7.2",
-    "eslint": "^6.5.1",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-react": "^7.0.0",
-    "jasmine": "^2.5.2",
+    "husky": "^4.2.5",
+    "jest": "^26.6.3",
+    "jest-date-mock": "^1.0.8",
+    "lint-staged": "^10.2.11",
+    "metro-react-native-babel-preset": "^0.64.0",
     "mocha": "^7.1.0",
+    "prettier": "^2.0.5",
     "react": "16.11.0",
+    "react-component-driver": "^0.10.0",
     "react-native": "0.62.2",
     "react-native-navigation": "^6.7.1",
+    "react-test-renderer": "^17.0.1",
     "semver": "5.x.x",
     "shell-utils": "1.x.x"
+  },
+  "jest": {
+    "preset": "react-native",
+    "setupFiles": [
+      "jest-date-mock",
+      "./jest-setup.js"
+    ],
+    "testMatch": [
+      "<rootDir>/src/**/?(*.)(spec|test).{js,jsx}"
+    ]
   },
   "detox": {
     "configurations": {
@@ -55,7 +79,8 @@
         "build": "xcodebuild -workspace ios/CalendarsExample.xcworkspace -scheme CalendarsExample -configuration Debug -sdk iphonesimulator -derivedDataPath ios/build",
         "type": "ios.simulator",
         "device": {
-          "type": "iPhone 11"
+          "type": "iPhone 11",
+          "os": "iOS 13.7"
         }
       },
       "ios.sim.release": {
@@ -63,10 +88,20 @@
         "build": "xcodebuild -workspace ios/CalendarsExample.xcworkspace -scheme CalendarsExample -configuration Release -sdk iphonesimulator -derivedDataPath ios/build",
         "type": "ios.simulator",
         "device": {
-          "type": "iPhone 11"
+          "type": "iPhone 11",
+          "os": "iOS 13.7"
         }
       }
     },
     "test-runner": "mocha"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.js": "eslint --cache --fix",
+    "*.{js,css,md}": "prettier --write"
   }
 }
