@@ -1,6 +1,8 @@
+import './wdyr'; // <--- must be first import
+
 import {Navigation} from 'react-native-navigation';
-import {registerScreens} from './screens';
 import {LocaleConfig} from 'react-native-calendars';
+import {registerScreens} from './screens';
 
 registerScreens();
 // eslint-disable-next-line no-console
@@ -17,7 +19,7 @@ LocaleConfig.locales['fr'] = {
 LocaleConfig.defaultLocale = 'fr';
 */
 LocaleConfig.locales['en'] = {
-  formatAccessibilityLabel: 'dddd d \'of\' MMMM \'of\' yyyy',
+  formatAccessibilityLabel: "dddd d 'of' MMMM 'of' yyyy",
   monthNames: [
     'January',
     'February',
@@ -32,26 +34,12 @@ LocaleConfig.locales['en'] = {
     'November',
     'December'
   ],
-  monthNamesShort: [
-    'jan',
-    'feb',
-    'mar',
-    'apr',
-    'may',
-    'jun',
-    'jul',
-    'aug',
-    'sep',
-    'oct',
-    'nov',
-    'dec'
-  ],
+  monthNamesShort: ['jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'],
   dayNames: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
   dayNamesShort: ['S', 'M', 'T', 'W', 'T', 'F', 'S']
 };
 
 LocaleConfig.defaultLocale = 'en';
-
 
 Navigation.events().registerAppLaunchedListener(() => {
   Navigation.setRoot({
