@@ -1,18 +1,9 @@
 import _ from 'lodash';
 import React, {Component} from 'react';
-import {
-  Platform,
-  Alert,
-  StyleSheet,
-  View,
-  Text,
-  TouchableOpacity,
-  Button
-} from 'react-native';
+import {Platform, Alert, StyleSheet, View, Text, TouchableOpacity, Button} from 'react-native';
 import {ExpandableCalendar, AgendaList, CalendarProvider, WeekCalendar} from 'react-native-calendars';
 
 const testIDs = require('../testIDs');
-
 
 const today = new Date().toISOString().split('T')[0];
 const fastDate = getPastDate(3);
@@ -24,7 +15,7 @@ const lightThemeColor = '#EBF9F9';
 function getFutureDates(days) {
   const array = [];
   for (let index = 1; index <= days; index++) {
-    const date = new Date(Date.now() + (864e5 * index)); // 864e5 == 86400000 == 24*60*60*1000
+    const date = new Date(Date.now() + 864e5 * index); // 864e5 == 86400000 == 24*60*60*1000
     const dateString = date.toISOString().split('T')[0];
     array.push(dateString);
   }
@@ -32,33 +23,68 @@ function getFutureDates(days) {
 }
 
 function getPastDate(days) {
-  return new Date(Date.now() - (864e5 * days)).toISOString().split('T')[0];
+  return new Date(Date.now() - 864e5 * days).toISOString().split('T')[0];
 }
 
 const ITEMS = [
   {title: dates[0], data: [{hour: '12am', duration: '1h', title: 'First Yoga'}]},
-  {title: dates[1], data: [{hour: '4pm', duration: '1h', title: 'Pilates ABC'}, {hour: '5pm', duration: '1h', title: 'Vinyasa Yoga'}]},
-  {title: dates[2], data: [{hour: '1pm', duration: '1h', title: 'Ashtanga Yoga'}, {hour: '2pm', duration: '1h', title: 'Deep Streches'}, {hour: '3pm', duration: '1h', title: 'Private Yoga'}]},
+  {
+    title: dates[1],
+    data: [
+      {hour: '4pm', duration: '1h', title: 'Pilates ABC'},
+      {hour: '5pm', duration: '1h', title: 'Vinyasa Yoga'}
+    ]
+  },
+  {
+    title: dates[2],
+    data: [
+      {hour: '1pm', duration: '1h', title: 'Ashtanga Yoga'},
+      {hour: '2pm', duration: '1h', title: 'Deep Streches'},
+      {hour: '3pm', duration: '1h', title: 'Private Yoga'}
+    ]
+  },
   {title: dates[3], data: [{hour: '12am', duration: '1h', title: 'Ashtanga Yoga'}]},
   {title: dates[4], data: [{}]},
-  {title: dates[5], data: [{hour: '9pm', duration: '1h', title: 'Middle Yoga'}, {hour: '10pm', duration: '1h', title: 'Ashtanga'}, {hour: '11pm', duration: '1h', title: 'TRX'}, {hour: '12pm', duration: '1h', title: 'Running Group'}]},
+  {
+    title: dates[5],
+    data: [
+      {hour: '9pm', duration: '1h', title: 'Middle Yoga'},
+      {hour: '10pm', duration: '1h', title: 'Ashtanga'},
+      {hour: '11pm', duration: '1h', title: 'TRX'},
+      {hour: '12pm', duration: '1h', title: 'Running Group'}
+    ]
+  },
   {title: dates[6], data: [{hour: '12am', duration: '1h', title: 'Ashtanga Yoga'}]},
   {title: dates[7], data: [{}]},
-  {title: dates[8], data: [{hour: '9pm', duration: '1h', title: 'Pilates Reformer'}, {hour: '10pm', duration: '1h', title: 'Ashtanga'}, {hour: '11pm', duration: '1h', title: 'TRX'}, {hour: '12pm', duration: '1h', title: 'Running Group'}]},
-  {title: dates[9], data: [{hour: '1pm', duration: '1h', title: 'Ashtanga Yoga'}, {hour: '2pm', duration: '1h', title: 'Deep Streches'}, {hour: '3pm', duration: '1h', title: 'Private Yoga'}]},
+  {
+    title: dates[8],
+    data: [
+      {hour: '9pm', duration: '1h', title: 'Pilates Reformer'},
+      {hour: '10pm', duration: '1h', title: 'Ashtanga'},
+      {hour: '11pm', duration: '1h', title: 'TRX'},
+      {hour: '12pm', duration: '1h', title: 'Running Group'}
+    ]
+  },
+  {
+    title: dates[9],
+    data: [
+      {hour: '1pm', duration: '1h', title: 'Ashtanga Yoga'},
+      {hour: '2pm', duration: '1h', title: 'Deep Streches'},
+      {hour: '3pm', duration: '1h', title: 'Private Yoga'}
+    ]
+  },
   {title: dates[10], data: [{hour: '12am', duration: '1h', title: 'Last Yoga'}]}
 ];
 
 export default class ExpandableCalendarScreen extends Component {
-
   onDateChanged = (/* date, updateSource */) => {
     // console.warn('ExpandableCalendarScreen onDateChanged: ', date, updateSource);
     // fetch and set data for date + week ahead
-  }
+  };
 
   onMonthChange = (/* month, updateSource */) => {
     // console.warn('ExpandableCalendarScreen onMonthChange: ', month, updateSource);
-  }
+  };
 
   buttonPressed() {
     Alert.alert('show more');
@@ -82,22 +108,18 @@ export default class ExpandableCalendarScreen extends Component {
     }
 
     return (
-      <TouchableOpacity
-        onPress={() => this.itemPressed(item.title)}
-        style={styles.item}
-        testID={testIDs.agenda.ITEM}
-      >
+      <TouchableOpacity onPress={() => this.itemPressed(item.title)} style={styles.item} testID={testIDs.agenda.ITEM}>
         <View>
           <Text style={styles.itemHourText}>{item.hour}</Text>
           <Text style={styles.itemDurationText}>{item.duration}</Text>
         </View>
         <Text style={styles.itemTitleText}>{item.title}</Text>
         <View style={styles.itemButtonContainer}>
-          <Button color={'grey'} title={'Info'} onPress={this.buttonPressed}/>
+          <Button color={'grey'} title={'Info'} onPress={this.buttonPressed} />
         </View>
       </TouchableOpacity>
     );
-  }
+  };
 
   getMarkedDates = () => {
     const marked = {};
@@ -110,7 +132,7 @@ export default class ExpandableCalendarScreen extends Component {
       }
     });
     return marked;
-  }
+  };
 
   getTheme = () => {
     const disabledColor = 'grey';
@@ -146,7 +168,7 @@ export default class ExpandableCalendarScreen extends Component {
       disabledDotColor: disabledColor,
       dotStyle: {marginTop: -2}
     };
-  }
+  };
 
   render() {
     return (
@@ -161,12 +183,9 @@ export default class ExpandableCalendarScreen extends Component {
         // }}
         // todayBottomMargin={16}
       >
-        {this.props.weekView ?
-          <WeekCalendar
-            testID={testIDs.weekCalendar.CONTAINER}
-            firstDay={1}
-            markedDates={this.getMarkedDates()}
-          /> :
+        {this.props.weekView ? (
+          <WeekCalendar testID={testIDs.weekCalendar.CONTAINER} firstDay={1} markedDates={this.getMarkedDates()} />
+        ) : (
           <ExpandableCalendar
             testID={testIDs.expandableCalendar.CONTAINER}
             // horizontal={false}
@@ -184,7 +203,7 @@ export default class ExpandableCalendarScreen extends Component {
             leftArrowImageSource={require('../img/previous.png')}
             rightArrowImageSource={require('../img/next.png')}
           />
-        }
+        )}
         <AgendaList
           sections={ITEMS}
           extraData={this.state}
