@@ -4,11 +4,10 @@ import {CalendarList} from 'react-native-calendars';
 const testIDs = require('../testIDs');
 
 const HorizontalCalendarList = () => {
-
   const [selectedDate, setSelectedDate] = React.useState('2020-05-16');
   const [markedDates, setMarkedDates] = React.useState({});
 
-  const setNewDaySelected = (date) => {
+  const setNewDaySelected = date => {
     const markedDate = Object.assign({});
     markedDate[date] = {
       selected: true,
@@ -27,7 +26,7 @@ const HorizontalCalendarList = () => {
       futureScrollRange={24}
       horizontal
       pagingEnabled
-      onDayPress={(day) => {
+      onDayPress={day => {
         setNewDaySelected(day.dateString);
       }}
     />
