@@ -3,9 +3,13 @@ import * as defaultStyle from '../../../style';
 
 const STYLESHEET_ID = 'stylesheet.day.basic';
 
-export default function styleConstructor(theme={}) {
+export default function styleConstructor(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
   return StyleSheet.create({
+    container: {
+      alignSelf: 'stretch',
+      alignItems: 'center'
+    },
     base: {
       width: 32,
       height: 32,
