@@ -1,15 +1,16 @@
 import {StyleSheet} from 'react-native';
-import * as defaultStyle from '../../style';
+import * as defaultStyle from '../../../style';
 
 const STYLESHEET_ID = 'stylesheet.dot';
 
-export default function styleConstructor(theme={}) {
+export default function styleConstructor(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
   return StyleSheet.create({
     dot: {
       width: 4,
       height: 4,
       marginTop: 1,
+      marginHorizontal: 1,
       borderRadius: 2,
       opacity: 0,
       ...appStyle.dotStyle
