@@ -1,110 +1,111 @@
-import React, {Component, Fragment} from 'react';
-import {ActivityIndicator, Platform} from 'react-native';
-import {View, Text, TouchableOpacity, Image} from 'react-native';
-import XDate from 'xdate';
-import PropTypes from 'prop-types';
-import styleConstructor from './style';
-import {weekDayNames} from '../../dateutils';
-import {CHANGE_MONTH_LEFT_ARROW, CHANGE_MONTH_RIGHT_ARROW, HEADER_MONTH_NAME} from '../../testIDs';
 import _ from 'lodash';
-
+import PropTypes from 'prop-types';
+import XDate from 'xdate';
+import React, {Component, Fragment} from 'react';
+import {ActivityIndicator, Platform, View, Text, TouchableOpacity, Image} from 'react-native';
+import {shouldUpdate} from '../../component-updater';
+import {weekDayNames} from '../../dateutils';
+import {
+  CHANGE_MONTH_LEFT_ARROW,
+  CHANGE_MONTH_RIGHT_ARROW,
+  HEADER_DAY_NAMES,
+  HEADER_LOADING_INDICATOR,
+  HEADER_MONTH_NAME
+} from '../../testIDs';
+import styleConstructor from './style';
 
 class CalendarHeader extends Component {
   static displayName = 'IGNORE';
 
   static propTypes = {
     theme: PropTypes.object,
-    hideArrows: PropTypes.bool,
+    firstDay: PropTypes.number,
+    displayLoadingIndicator: PropTypes.bool,
+    showWeekNumbers: PropTypes.bool,
     month: PropTypes.instanceOf(XDate),
     addMonth: PropTypes.func,
-    showIndicator: PropTypes.bool,
-    firstDay: PropTypes.number,
-    renderArrow: PropTypes.func,
+    /** Month format in the title. Formatting values: http://arshaw.com/xdate/#Formatting */
+    monthFormat: PropTypes.string,
+    /**  Hide day names. Default = false */
     hideDayNames: PropTypes.bool,
-    weekNumbers: PropTypes.bool,
+    /** Hide month navigation arrows. Default = false */
+    hideArrows: PropTypes.bool,
+    /** Replace default arrows with custom ones (direction can be 'left' or 'right') */
+    renderArrow: PropTypes.func,
+    /** Handler which gets executed when press arrow icon left. It receive a callback can go back month */
     onPressArrowLeft: PropTypes.func,
+    /** Handler which gets executed when press arrow icon right. It receive a callback can go next month */
     onPressArrowRight: PropTypes.func,
+    /** Disable left arrow. Default = false */
     disableArrowLeft: PropTypes.bool,
+    /** Disable right arrow. Default = false */
     disableArrowRight: PropTypes.bool,
-    webAriaLevel: PropTypes.number,
+    /** Apply custom disable color to selected day indexes */
     disabledDaysIndexes: PropTypes.arrayOf(PropTypes.number),
+    /** Replace default month and year title with custom one. the function receive a date as parameter. */
     renderHeader: PropTypes.any,
-    renderTopHeader: PropTypes.bool
+    /** Provide aria-level for calendar heading for proper accessibility when used with web (react-native-web) */
+    webAriaLevel: PropTypes.number
   };
 
   static defaultProps = {
     monthFormat: 'MMMM yyyy',
-    webAriaLevel: 1,
-    renderTopHeader: true,
+    webAriaLevel: 1
   };
 
   constructor(props) {
     super(props);
+
     this.style = styleConstructor(props.theme);
-  }
-
-  addMonth = () => {
-    const {addMonth} = this.props;
-    addMonth(1);
-  }
-
-  subtractMonth = () => {
-    const {addMonth} = this.props;
-    addMonth(-1);
   }
 
   shouldComponentUpdate(nextProps) {
     if (nextProps.month.toString('yyyy MM') !== this.props.month.toString('yyyy MM')) {
       return true;
     }
-    if (nextProps.showIndicator !== this.props.showIndicator) {
-      return true;
-    }
-    if (nextProps.hideDayNames !== this.props.hideDayNames) {
-      return true;
-    }
-    if (nextProps.firstDay !== this.props.firstDay) {
-      return true;
-    }
-    if (nextProps.weekNumbers !== this.props.weekNumbers) {
-      return true;
-    }
-    if (nextProps.monthFormat !== this.props.monthFormat) {
-      return true;
-    }
-    if (nextProps.renderArrow !== this.props.renderArrow) {
-      return true;
-    }
-    if (nextProps.disableArrowLeft !== this.props.disableArrowLeft) {
-      return true;
-    }
-    if (nextProps.disableArrowRight !== this.props.disableArrowRight) {
-      return true;
-    }
-    if (nextProps.renderTopHeader !== this.props.renderTopHeader) {
-      return true;
-    }
-    return false;
+    return shouldUpdate(this.props, nextProps, [
+      'displayLoadingIndicator',
+      'hideDayNames',
+      'firstDay',
+      'showWeekNumbers',
+      'monthFormat',
+      'renderArrow',
+      'disableArrowLeft',
+      'disableArrowRight'
+    ]);
   }
+
+  addMonth = () => {
+    const {addMonth} = this.props;
+    addMonth(1);
+  };
+
+  subtractMonth = () => {
+    const {addMonth} = this.props;
+    addMonth(-1);
+  };
 
   onPressLeft = () => {
     const {onPressArrowLeft, month} = this.props;
+
     if (typeof onPressArrowLeft === 'function') {
       return onPressArrowLeft(this.subtractMonth, month);
     }
     return this.subtractMonth();
-  }
+  };
 
   onPressRight = () => {
     const {onPressArrowRight, month} = this.props;
+
     if (typeof onPressArrowRight === 'function') {
       return onPressArrowRight(this.addMonth, month);
     }
     return this.addMonth();
-  }
+  };
 
-  renderWeekDays = (weekDaysNames) => {
+  renderWeekDays = weekDaysNames => {
     const {disabledDaysIndexes} = this.props;
+
     return weekDaysNames.map((day, idx) => {
       const dayStyle = [this.style.dayHeader];
 
@@ -113,22 +114,16 @@ class CalendarHeader extends Component {
       }
 
       return (
-        <Text
-          allowFontScaling={false}
-          key={idx}
-          style={dayStyle}
-          numberOfLines={1}
-          accessibilityLabel={''}
-        >
+        <Text allowFontScaling={false} key={idx} style={dayStyle} numberOfLines={1} accessibilityLabel={''}>
           {day}
         </Text>
       );
     });
-  }
+  };
 
   renderHeader = () => {
-    const {renderHeader, month, monthFormat, testID} = this.props;
-    const webProps = Platform.OS === 'web' ? {'aria-level': this.props.webAriaLevel} : {};
+    const {renderHeader, month, monthFormat, testID, webAriaLevel} = this.props;
+    const webProps = Platform.OS === 'web' ? {'aria-level': webAriaLevel} : {};
 
     if (renderHeader) {
       return renderHeader(month);
@@ -139,7 +134,7 @@ class CalendarHeader extends Component {
         <Text
           allowFontScaling={false}
           style={this.style.monthText}
-          testID={testID ? `${HEADER_MONTH_NAME}-${testID}`: HEADER_MONTH_NAME}
+          testID={testID ? `${HEADER_MONTH_NAME}-${testID}` : HEADER_MONTH_NAME}
           {...webProps}
         >
           {month.toString(monthFormat)}
@@ -148,56 +143,70 @@ class CalendarHeader extends Component {
     );
   };
 
+  renderArrow(direction) {
+    const {hideArrows, disableArrowLeft, disableArrowRight, renderArrow, testID} = this.props;
+    if (hideArrows) {
+      return <View />;
+    }
+    const isLeft = direction === 'left';
+    const id = isLeft ? CHANGE_MONTH_LEFT_ARROW : CHANGE_MONTH_RIGHT_ARROW;
+    const testId = testID ? `${id}-${testID}` : id;
+    const onPress = isLeft ? this.onPressLeft : this.onPressRight;
+    const imageSource = isLeft ? require('../img/previous.png') : require('../img/next.png');
+    const renderArrowDirection = isLeft ? 'left' : 'right';
+    const shouldDisable = isLeft ? disableArrowLeft : disableArrowRight;
+
+    return (
+      <TouchableOpacity
+        onPress={!shouldDisable ? onPress : undefined}
+        disabled={shouldDisable}
+        style={this.style.arrow}
+        hitSlop={{left: 20, right: 20, top: 20, bottom: 20}}
+        testID={testId}
+      >
+        {renderArrow ? (
+          renderArrow(renderArrowDirection)
+        ) : (
+          <Image source={imageSource} style={shouldDisable ? this.style.disabledArrowImage : this.style.arrowImage} />
+        )}
+      </TouchableOpacity>
+    );
+  }
+
+  renderIndicator() {
+    const {displayLoadingIndicator, theme, testID} = this.props;
+
+    if (displayLoadingIndicator) {
+      return (
+        <ActivityIndicator
+          color={theme && theme.indicatorColor}
+          testID={testID ? `${HEADER_LOADING_INDICATOR}-${testID}` : HEADER_LOADING_INDICATOR}
+        />
+      );
+    }
+  }
+
+  renderDayNames() {
+    const {firstDay, hideDayNames, showWeekNumbers, testID} = this.props;
+    const weekDaysNames = weekDayNames(firstDay);
+
+    if (!hideDayNames) {
+      return (
+        <View style={this.style.week} testID={testID ? `${HEADER_DAY_NAMES}-${testID}` : HEADER_DAY_NAMES}>
+          {showWeekNumbers && <Text allowFontScaling={false} style={this.style.dayHeader}></Text>}
+          {this.renderWeekDays(weekDaysNames)}
+        </View>
+      );
+    }
+  }
+
   render() {
-    let leftArrow = <View/>;
-    let rightArrow = <View/>;
-    let weekDaysNames = weekDayNames(this.props.firstDay);
-    const {testID} = this.props;
-
-    if (!this.props.hideArrows) {
-      leftArrow = (
-        <TouchableOpacity
-          onPress={this.onPressLeft}
-          disabled={this.props.disableArrowLeft}
-          style={this.style.arrow}
-          hitSlop={{left: 20, right: 20, top: 20, bottom: 20}}
-          testID={testID ? `${CHANGE_MONTH_LEFT_ARROW}-${testID}`: CHANGE_MONTH_LEFT_ARROW}
-        >
-          {this.props.renderArrow
-            ? this.props.renderArrow('left')
-            : <Image
-              source={require('../img/previous.png')}
-              style={this.props.disableArrowLeft ? this.style.disabledArrowImage : this.style.arrowImage}
-            />}
-        </TouchableOpacity>
-      );
-      rightArrow = (
-        <TouchableOpacity
-          onPress={this.onPressRight}
-          disabled={this.props.disableArrowRight}
-          style={this.style.arrow}
-          hitSlop={{left: 20, right: 20, top: 20, bottom: 20}}
-          testID={testID ? `${CHANGE_MONTH_RIGHT_ARROW}-${testID}`: CHANGE_MONTH_RIGHT_ARROW}
-        >
-          {this.props.renderArrow
-            ? this.props.renderArrow('right')
-            : <Image
-              source={require('../img/next.png')}
-              style={this.props.disableArrowRight ? this.style.disabledArrowImage : this.style.arrowImage}
-            />}
-        </TouchableOpacity>
-      );
-    }
-
-    let indicator;
-    if (this.props.showIndicator) {
-      indicator = <ActivityIndicator color={this.props.theme && this.props.theme.indicatorColor}/>;
-    }
+    const {style, testID} = this.props;
 
     return (
       <View
         testID={testID}
-        style={this.props.style}
+        style={style}
         accessible
         accessibilityRole={'adjustable'}
         accessibilityActions={[
@@ -208,40 +217,31 @@ class CalendarHeader extends Component {
         accessibilityElementsHidden={this.props.accessibilityElementsHidden} // iOS
         importantForAccessibility={this.props.importantForAccessibility} // Android
       >
-        {this.props.renderTopHeader &&
         <View style={this.style.header}>
-          {leftArrow}
+          {this.renderArrow('left')}
           <View style={this.style.headerContainer}>
             {this.renderHeader()}
-            {indicator}
+            {this.renderIndicator()}
           </View>
-          {rightArrow}
+          {this.renderArrow('right')}
         </View>
-        }
-        {!this.props.hideDayNames &&
-          <View style={this.style.week}>
-            {this.props.weekNumbers &&
-              <Text allowFontScaling={false} style={this.style.dayHeader}></Text>
-            }
-            {this.renderWeekDays(weekDaysNames)}
-          </View>
-        }
+        {this.renderDayNames()}
       </View>
     );
   }
 
   onAccessibilityAction = event => {
     switch (event.nativeEvent.actionName) {
-    case 'decrement':
-      this.onPressLeft();
-      break;
-    case 'increment':
-      this.onPressRight();
-      break;
-    default:
-      break;
+      case 'decrement':
+        this.onPressLeft();
+        break;
+      case 'increment':
+        this.onPressRight();
+        break;
+      default:
+        break;
     }
-  }
+  };
 }
 
 export default CalendarHeader;
