@@ -3,7 +3,8 @@ import React, {Component} from 'react';
 import {SectionList, Text} from 'react-native';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
-
+import moment from 'moment';
+import dateutils from '../dateutils';
 import styleConstructor from './style';
 import asCalendarConsumer from './asCalendarConsumer';
 
@@ -13,8 +14,8 @@ const UPDATE_SOURCES = commons.UPDATE_SOURCES;
 
 /**
  * @description: AgendaList component
+ * @note: Should be wrapped with 'CalendarProvider'
  * @extends: SectionList
- * @notes: Should be wraped in CalendarProvider component
  * @example: https://github.com/wix/react-native-calendars/blob/master/example/src/screens/expandableCalendar.js
  */
 class AgendaList extends Component {
@@ -24,13 +25,21 @@ class AgendaList extends Component {
     ...SectionList.propTypes,
     /** day format in section title. Formatting values: http://arshaw.com/xdate/#Formatting */
     dayFormat: PropTypes.string,
+    /** a function to custom format the section header's title */
+    dayFormatter: PropTypes.func,
+    /** whether to use moment.js for date string formatting 
+     * (remember to pass 'dayFormat' with appropriate format, like 'dddd, MMM D') */
+    useMoment: PropTypes.bool,
+    /** whether to mark today's title with the "Today, ..." string. Default = true */
+    markToday: PropTypes.bool,
     /** style passed to the section view */
     sectionStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.number, PropTypes.array])
   }
 
   static defaultProps = {
     dayFormat: 'dddd, MMM d',
-    stickySectionHeadersEnabled: true
+    stickySectionHeadersEnabled: true,
+    markToday: true
   }
 
   constructor(props) {
@@ -131,23 +140,44 @@ class AgendaList extends Component {
   }
 
   renderSectionHeader = ({section: {title}}) => {
+    const {renderSectionHeader, dayFormatter, dayFormat, useMoment, markToday, sectionStyle} = this.props;
+
+    if (renderSectionHeader) {
+      return renderSectionHeader(title);
+    }
+
     let sectionTitle = title;
 
-    if (this.props.dayFormat) {
-      const date = XDate(title).toString(this.props.dayFormat);
-      const today = XDate().toString(this.props.dayFormat);
+    if (dayFormatter) {
+      sectionTitle = dayFormatter(title);
+    } else if (dayFormat) {
+      if (useMoment) {
+        sectionTitle = moment(title).format(dayFormat);
+      } else {
+        sectionTitle = XDate(title).toString(dayFormat);
+      }
+    }
+
+    if (markToday) {
       const todayString = XDate.locales[XDate.defaultLocale].today || commons.todayString;
-      
-      sectionTitle = date === today ? `${todayString}, ${date}` : date;
+      const isToday = dateutils.sameDate(XDate(), XDate(title));
+      sectionTitle = isToday ? `${todayString}, ${sectionTitle}` : sectionTitle;
     }
 
     return (
-      <Text allowFontScaling={false} style={[this.style.sectionText, this.props.sectionStyle]} onLayout={this.onHeaderLayout}>{sectionTitle}</Text>
+      <Text 
+        allowFontScaling={false} 
+        style={[this.style.sectionText, sectionStyle]} 
+        onLayout={this.onHeaderLayout}
+      >
+        {sectionTitle}
+      </Text>
     );
   }
 
   keyExtractor = (item, index) => {
-    return _.isFunction(this.props.keyExtractor) ? this.props.keyExtractor(item, index) : String(index);
+    const {keyExtractor} = this.props;
+    return _.isFunction(keyExtractor) ? keyExtractor(item, index) : String(index);
   }
 
   render() {
