@@ -1,19 +1,14 @@
-import React, {Component} from 'react';
-import {View} from 'react-native';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
-
+import React, {Component} from 'react';
+import {View} from 'react-native';
 import dateutils from '../dateutils';
-import {xdateToData, parseDate} from '../interface';
-import {SELECT_DATE_SLOT} from '../testIDs';
+import {parseDate} from '../interface';
+import {extractComponentProps} from '../component-updater';
 import styleConstructor from './style';
-
-import Day from '../calendar/day/basic';
-import UnitDay from '../calendar/day/period';
-import MultiDotDay from '../calendar/day/multi-dot';
-import MultiPeriodDay from '../calendar/day/multi-period';
-import SingleDay from '../calendar/day/custom';
 import Calendar from '../calendar';
+import Day from '../calendar/day/index';
+// import BasicDay from '../calendar/day/basic';
 
 
 const EmptyArray = [];
@@ -23,7 +18,7 @@ class Week extends Component {
 
   static propTypes = {
     ...Calendar.propTypes,
-    // the current date
+    /** the current date */
     current: PropTypes.any
   };
 
@@ -57,27 +52,8 @@ class Week extends Component {
         daysArray.push(newDate);
         index += 1;
       }
+
       return daysArray;
-    }
-  }
-
-  getDayComponent() {
-    const {dayComponent} = this.props;
-    if (dayComponent) {
-      return dayComponent;
-    }
-
-    switch (this.props.markingType) {
-    case 'period':
-      return UnitDay;
-    case 'multi-dot':
-      return MultiDotDay;
-    case 'multi-period':
-      return MultiPeriodDay;
-    case 'custom':
-      return SingleDay;
-    default:
-      return Day;
     }
   }
 
@@ -95,51 +71,49 @@ class Week extends Component {
     }
   }
 
-  // renderWeekNumber (weekNumber) {
-  //   return <Day key={`week-${weekNumber}`} theme={this.props.theme} marking={{disableTouchEvent: true}} state='disabled'>{weekNumber}</Day>;
-  // }
-
-  renderDay(day, id) {
-    const {current, disableAllTouchEventsForDisabledDays} = this.props;
+  getState(day) {
+    const {current, disabledByDefault} = this.props;
     const minDate = parseDate(this.props.minDate);
     const maxDate = parseDate(this.props.maxDate);
-
     let state = '';
-    if (this.props.disabledByDefault) {
+
+    if (disabledByDefault) {
       state = 'disabled';
-    } else if ((minDate && !dateutils.isGTE(day, minDate)) || (maxDate && !dateutils.isLTE(day, maxDate))) {
+    } else if (dateutils.isDateNotInTheRange(minDate, maxDate, day)) {
       state = 'disabled';
-    } else if (!dateutils.sameMonth(day, parseDate(current))) { // for extra days
+    } else if (!dateutils.sameMonth(day, parseDate(current))) {
       state = 'disabled';
     } else if (dateutils.sameDate(day, XDate())) {
       state = 'today';
     }
+    return state;
+  }
+
+  // renderWeekNumber (weekNumber) {
+  //   return <BasicDay key={`week-${weekNumber}`} theme={this.props.theme} marking={{disableTouchEvent: true}} state='disabled'>{weekNumber}</Day>;
+  // }
+  
+  renderDay(day, id) {
+    const {current, hideExtraDays} = this.props;
+    const dayProps = extractComponentProps(Day, this.props);
 
     // hide extra days
-    if (current && this.props.hideExtraDays) {
+    if (current && hideExtraDays) {
       if (!dateutils.sameMonth(day, parseDate(current))) {
-        return (<View key={id} style={{flex: 1}}/>);
+        return (<View key={id} style={this.style.emptyDayContainer}/>);
       }
     }
 
-    const DayComp = this.getDayComponent();
-    const dayDate = day.getDate();
-    const dateAsObject = xdateToData(day);
-
     return (
-      <View style={{flex: 1, alignItems: 'center'}} key={id}>
-        <DayComp
-          testID={`${SELECT_DATE_SLOT}-${dateAsObject.dateString}`}
-          state={state}
-          theme={this.props.theme}
+      <View style={this.style.dayContainer} key={id}>
+        <Day
+          {...dayProps}
+          day={day}
+          state={this.getState(day)}
+          marking={this.getDateMarking(day)}
           onPress={this.props.onDayPress}
           onLongPress={this.props.onDayPress}
-          date={dateAsObject}
-          marking={this.getDateMarking(day)}
-          disableAllTouchEventsForDisabledDays={disableAllTouchEventsForDisabledDays}
-        >
-          {dayDate}
-        </DayComp>
+        />
       </View>
     );
   }
