@@ -8,5 +8,8 @@ module.exports = {
   STATIC_HEADER: 'STATIC_HEADER',
   AGENDA_CALENDAR_KNOB: `${PREFIX}.AGENDA_CALENDAR_KNOB`,
   HEADER_MONTH_NAME: 'HEADER_MONTH_NAME',
-  RESERVATION_DATE: `${PREFIX}.RESERVATION_DATE`
+  RESERVATION_DATE: `${PREFIX}.RESERVATION_DATE`,
+  HEADER_DAY_NAMES: `${PREFIX}.DAY_NAMES`,
+  WEEK_NUMBER: `${PREFIX}.WEEK_NUMBER`,
+  HEADER_LOADING_INDICATOR: `${PREFIX}.HEADER_LOADING_INDICATOR`
 };
