@@ -1,17 +1,11 @@
 // @flow
-import {
-  View,
-  Text,
-  ScrollView,
-  TouchableOpacity,
-  Dimensions
-} from 'react-native';
-import PropTypes from 'prop-types';
-import populateEvents from './Packer';
-import React from 'react';
-import moment from 'moment';
 import _ from 'lodash';
+import PropTypes from 'prop-types';
+import moment from 'moment';
+import React from 'react';
+import {View, Text, ScrollView, TouchableOpacity, Dimensions} from 'react-native';
 import styleConstructor from './style';
+import populateEvents from './Packer';
 
 const LEFT_MARGIN = 60 - 1;
 const TEXT_LINE_HEIGHT = 17;
@@ -28,32 +22,37 @@ export default class Timeline extends React.PureComponent {
     end: PropTypes.number,
     eventTapped: PropTypes.func,
     format24h: PropTypes.bool,
-    events: PropTypes.arrayOf(PropTypes.shape({
-      start: PropTypes.string.isRequired,
-      end: PropTypes.string.isRequired,
-      title: PropTypes.string.isRequired,
-      summary: PropTypes.string.isRequired,
-      color: PropTypes.string
-    })).isRequired
-  }
+    events: PropTypes.arrayOf(
+      PropTypes.shape({
+        start: PropTypes.string.isRequired,
+        end: PropTypes.string.isRequired,
+        title: PropTypes.string.isRequired,
+        summary: PropTypes.string.isRequired,
+        color: PropTypes.string
+      })
+    ).isRequired
+  };
 
   static defaultProps = {
     start: 0,
     end: 24,
     events: [],
     format24h: true
-  }
+  };
 
   constructor(props) {
     super(props);
+
     const {start, end} = this.props;
     this.calendarHeight = (end - start) * 100;
-    this.styles = styleConstructor(props.styles, this.calendarHeight);
+
+    this.style = styleConstructor(props.styles, this.calendarHeight);
+
     const width = dimensionWidth - LEFT_MARGIN;
     const packedEvents = populateEvents(props.events, width, start);
-    let initPosition =
-      _.min(_.map(packedEvents, 'top')) - this.calendarHeight / (end - start);
+    let initPosition = _.min(_.map(packedEvents, 'top')) - this.calendarHeight / (end - start);
     const verifiedInitPosition = initPosition < 0 ? 0 : initPosition;
+
     this.state = {
       _scrollY: verifiedInitPosition,
       packedEvents
@@ -64,7 +63,8 @@ export default class Timeline extends React.PureComponent {
     const width = dimensionWidth - LEFT_MARGIN;
     const {events: prevEvents, start: prevStart = 0} = prevProps;
     const {events, start = 0} = this.props;
-    if(prevEvents !== events || prevStart !== start) {
+
+    if (prevEvents !== events || prevStart !== start) {
       this.setState({
         packedEvents: populateEvents(events, width, start)
       });
@@ -90,11 +90,11 @@ export default class Timeline extends React.PureComponent {
   _renderLines() {
     const {format24h, start = 0, end = 24} = this.props;
     const offset = this.calendarHeight / (end - start);
-
     const EVENT_DIFF = 20;
 
     return range(start, end + 1).map((i, index) => {
       let timeText;
+      
       if (i === start) {
         timeText = '';
       } else if (i < 12) {
@@ -106,34 +106,24 @@ export default class Timeline extends React.PureComponent {
       } else {
         timeText = !format24h ? `${i - 12} PM` : `${i}:00`;
       }
+
       return [
-        <Text
-          key={`timeLabel${i}`}
-          style={[this.styles.timeLabel, {top: offset * index - 6}]}>
+        <Text key={`timeLabel${i}`} style={[this.style.timeLabel, {top: offset * index - 6}]}>
           {timeText}
         </Text>,
         i === start ? null : (
-          <View
-            key={`line${i}`}
-            style={[
-              this.styles.line,
-              {top: offset * index, width: dimensionWidth - EVENT_DIFF}
-            ]}
-          />
+          <View key={`line${i}`} style={[this.style.line, {top: offset * index, width: dimensionWidth - EVENT_DIFF}]} />
         ),
         <View
           key={`lineHalf${i}`}
-          style={[
-            this.styles.line,
-            {top: offset * (index + 0.5), width: dimensionWidth - EVENT_DIFF}
-          ]}
+          style={[this.style.line, {top: offset * (index + 0.5), width: dimensionWidth - EVENT_DIFF}]}
         />
       ];
     });
   }
 
   _onEventTapped(event) {
-    if(this.props.eventTapped) {
+    if (this.props.eventTapped) {
       this.props.eventTapped(event);
     }
   }
@@ -153,30 +143,29 @@ export default class Timeline extends React.PureComponent {
       // However it would make sense to overflow the title to a new line if needed
       const numberOfLines = Math.floor(event.height / TEXT_LINE_HEIGHT);
       const formatTime = this.props.format24h ? 'HH:mm' : 'hh:mm A';
+      
       return (
         <TouchableOpacity
           activeOpacity={0.9}
           onPress={() => this._onEventTapped(this.props.events[event.index])}
           key={i}
-          style={[this.styles.event, style]}>
+          style={[this.style.event, style]}
+        >
           {this.props.renderEvent ? (
             this.props.renderEvent(event)
           ) : (
             <View>
-              <Text numberOfLines={1} style={this.styles.eventTitle}>
+              <Text numberOfLines={1} style={this.style.eventTitle}>
                 {event.title || 'Event'}
               </Text>
               {numberOfLines > 1 ? (
-                <Text
-                  numberOfLines={numberOfLines - 1}
-                  style={[this.styles.eventSummary]}>
+                <Text numberOfLines={numberOfLines - 1} style={[this.style.eventSummary]}>
                   {event.summary || ' '}
                 </Text>
               ) : null}
               {numberOfLines > 2 ? (
-                <Text style={this.styles.eventTimes} numberOfLines={1}>
-                  {moment(event.start).format(formatTime)} -{' '}
-                  {moment(event.end).format(formatTime)}
+                <Text style={this.style.eventTimes} numberOfLines={1}>
+                  {moment(event.start).format(formatTime)} - {moment(event.end).format(formatTime)}
                 </Text>
               ) : null}
             </View>
@@ -196,10 +185,8 @@ export default class Timeline extends React.PureComponent {
     return (
       <ScrollView
         ref={ref => (this._scrollView = ref)}
-        contentContainerStyle={[
-          this.styles.contentStyle,
-          {width: dimensionWidth}
-        ]}>
+        contentContainerStyle={[this.style.contentStyle, {width: dimensionWidth}]}
+      >
         {this._renderLines()}
         {this._renderEvents()}
       </ScrollView>
